+++ DoorAccessIO/src/Face/Person/Parameter/ReadPersonDetail_Parameter.java
@@ -11,13 +11,13 @@ public class ReadPersonDetail_Parameter extends CommandParameter {
     /**
      * 用户号
      */
-    public int UserCode;
+    public long UserCode;
     /**
      * 查询人员数据详情
      * @param detail 命令详情
      * @param userCode 用户号
      */
-    public ReadPersonDetail_Parameter(CommandDetail detail,int userCode) {
+    public ReadPersonDetail_Parameter(CommandDetail detail,long userCode) {
         super(detail);
         UserCode=userCode;
     }
